# src/specs/api.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
  description: The main Node.js pipeline API.
  license:
    name: MIT
    url: 'https://github.com/hms-dbmi-cellenics/api/blob/master/LICENSE'
  contact:
    name: Biomage Ltd.
servers:
  - url: /v1
...
paths:
...
  '/experiments/{experimentId}/samples':
    parameters:
      - schema:
          type: string
        name: experimentId
        in: path
        required: true
    get:
      summary: Your GET endpoint
      operationId: getSamplesByExperimentId
      x-eov-operation-id: samples#getSamplesByExperimentId
      x-eov-operation-handler: routes/samples
      tags: []
      responses:
        '200':
          description: 'Fetch successful, samples returned below.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Samples'
        '401':
          description: The request lacks authentication credentials.
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
        '403':
          description: The authenticated user is not authorized to view this resource.
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
        '404':
          description: Samples not found.
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
      description: Get all samples for an experiment
...
  '/projects/{projectUuid}/{experimentId}/samples':
    parameters:
      - name: projectUuid
        in: path
        required: true
        schema:
          type: string
      - name: experimentId
        in: path
        required: true
        schema:
          type: string
    put:
      summary: Update samples
      operationId: updateSamples
      x-eov-operation-id: samples#update
      x-eov-operation-handler: routes/samples
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: ./models/HTTPSuccess.v1.yaml
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
        '415':
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
      description: Updates a samples entry
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Samples'
    post:
      summary: Add sample
      description: Adds a new sample to a samples entry
      operationId: addSample
      x-eov-operation-id: samples#add
      x-eov-operation-handler: routes/samples
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: ./models/HTTPSuccess.v1.yaml
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
        '415':
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sample'
    delete:
      summary: Remove samples
      operationId: removeSamples
      x-eov-operation-id: samples#remove
      x-eov-operation-handler: routes/samples
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: ./models/HTTPSuccess.v1.yaml
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
        '415':
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
      description: Removes samples from an entry
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SamplesRemoveBody'
  '/projects/{projectUuid}/samples/{sampleUuid}/{fileName}/uploadUrl':
    get:
      summary: Get file upload url
      description: Get temporary s3 upload url for fileName
      operationId: uploadSampleFileUrl
      x-eov-operation-id: samples#uploadSampleFileUrl
      x-eov-operation-handler: routes/samples
      parameters:
        - in: query
          name: cellranger_version
          description: The cellranger version of the dataset
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
  '/projects/{projectUuid}/samples/{sampleUuid}/{fileName}/downloadUrl':
    get:
      summary: Get file download url
      description: Get temporary s3 download url for fileName
      operationId: downloadSampleFileUrl
      x-eov-operation-id: samples#downloadSampleFileUrl
      x-eov-operation-handler: routes/samples
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
        '401':
          description: The request lacks authentication credentials.
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
        '403':
          description: The authenticated user is not authorized to view this resource.
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
        '404':
          description: Experiment not found
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
...
components:
  schemas:
...
    Samples:
      $ref: './models/api-body-schemas/Samples.v1.yaml'
    Sample:
      $ref: './models/samples-bodies/Sample.v1.yaml'
    SamplesRemoveBody:
      $ref: './models/samples-bodies/SamplesRemoveBody.v1.yaml'
...